+++ lib/datasource/remote/api/balance_repository.dart
@@ -1,6 +1,6 @@
 import 'package:async/async.dart';
 import 'package:http/http.dart' as http;
-import 'package:seeds/datasource/remote/api/network_repository.dart';
+import 'package:seeds/datasource/remote/api/http_repo/network_repository.dart';
 import 'package:seeds/datasource/remote/model/balance_model.dart';
 
 class BalanceRepository extends NetworkRepository {
